Simplify prepend_branch file rename with pre_hooks. Extract job commands to their own files. Bug fix: check if translation_file_ext is null.
